move sidebar functions to utils
